+++ src/App.js
@@ -2,8 +2,8 @@ import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Mypage from "./pages/users/mypage/index";
 import Main from "./pages/index";
-import SignUp from "./pages/users/SignUpPage";
-import SignIn from "./pages/users/SignIn";
+import SignUp from "./pages/users/signUp";
+import SignIn from "./pages/users/signIn";
 import CreateSurvey from "./pages/surveys/create/index";
 import routes from "./routes";
 import { createGlobalStyle } from "styled-components";
